requirements.txt
# Core dependencies
PyYAML>=6.0
pynput>=1.7.6
PyAudio>=0.2.13
faster-whisper>=0.10.0
python-xlib>=0.33
evdev>=1.6.1
PyGObject>=3.46.0
numpy>=1.24.0

# Optional but recommended
noisereduce>=3.0.0
watchdog>=3.0.0  # For live config file reloading

# Development dependencies (optional)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0
